Fix compile error under gcc.

+++ src/mission_ui.cpp
@@ -127,7 +127,7 @@ void game::list_missions()
             auto format_tokenized_description = []( const std::string description,
             std::vector<std::pair<int, std::string>> rewards ) {
                 std::string formatted_description = description;
-                for( int i = 0; i < rewards.size(); i++ ) {
+                for( size_t i = 0; i < rewards.size(); i++ ) {
                     std::regex token("<reward_count:" + rewards[i].second + ">");
                     formatted_description = std::regex_replace( formatted_description, token, string_format( "%d",
                                             rewards[i].first ) );
